--- web/packages/shared/storage/cacheStorage/constants.ts
@@ -1,6 +1,0 @@
-export enum CACHES {
-    THUMBS = "thumbs",
-    FACE_CROPS = "face-crops",
-    // Desktop app only
-    FILES = "files",
-}
